Moved setup methods
