+++ node/components/network/src/gossip/tests/fetch_blocks.rs
@@ -16,7 +16,7 @@ async fn test_simple() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = validator::testonly::Setup::new(rng, 1);
-    setup.push_blocks(rng, 2);
+    setup.push_blocks_v1(rng, 2);
     let mut cfg = crate::testonly::new_configs(rng, &setup, 0)[0].clone();
     cfg.rpc.push_block_store_state_rate = limiter::Rate::INF;
     cfg.rpc.get_block_rate = limiter::Rate::INF;
@@ -124,7 +124,7 @@ async fn test_concurrent_requests() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = validator::testonly::Setup::new(rng, 1);
-    setup.push_blocks(rng, 10);
+    setup.push_blocks_v1(rng, 10);
     let mut cfg = crate::testonly::new_configs(rng, &setup, 0)[0].clone();
     cfg.rpc.push_block_store_state_rate = limiter::Rate::INF;
     cfg.rpc.get_block_rate = limiter::Rate::INF;
@@ -199,7 +199,7 @@ async fn test_bad_responses() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = validator::testonly::Setup::new(rng, 1);
-    setup.push_blocks(rng, 2);
+    setup.push_blocks_v1(rng, 2);
     let mut cfg = crate::testonly::new_configs(rng, &setup, 0)[0].clone();
     cfg.rpc.push_block_store_state_rate = limiter::Rate::INF;
     cfg.rpc.get_block_rate = limiter::Rate::INF;
@@ -276,7 +276,7 @@ async fn test_retry() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = validator::testonly::Setup::new(rng, 1);
-    setup.push_blocks(rng, 1);
+    setup.push_blocks_v1(rng, 1);
     let mut cfg = crate::testonly::new_configs(rng, &setup, 0)[0].clone();
     cfg.rpc.push_block_store_state_rate = limiter::Rate::INF;
     cfg.rpc.get_block_rate = limiter::Rate::INF;
@@ -340,7 +340,7 @@ async fn test_announce_truncated_block_range() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = validator::testonly::Setup::new(rng, 1);
-    setup.push_blocks(rng, 10);
+    setup.push_blocks_v1(rng, 10);
     let mut cfg = crate::testonly::new_configs(rng, &setup, 0)[0].clone();
     cfg.rpc.push_block_store_state_rate = limiter::Rate::INF;
     cfg.rpc.get_block_rate = limiter::Rate::INF;
