+++ node/components/network/src/gossip/tests/syncing.rs
@@ -31,7 +31,7 @@ async fn coordinated_block_syncing(node_count: usize, gossip_peers: usize) {
     let mut spec = validator::testonly::SetupSpec::new(rng, node_count);
     spec.first_block = spec.first_pregenesis_block + 2;
     let mut setup = validator::testonly::Setup::from_spec(rng, spec);
-    setup.push_blocks(rng, EXCHANGED_STATE_COUNT);
+    setup.push_blocks_v1(rng, EXCHANGED_STATE_COUNT);
     let cfgs = testonly::new_configs(rng, &setup, gossip_peers);
     scope::run!(ctx, |ctx, s| async {
         let mut nodes = vec![];
@@ -89,7 +89,7 @@ async fn uncoordinated_block_syncing(
     let mut spec = validator::testonly::SetupSpec::new(rng, node_count);
     spec.first_block = spec.first_pregenesis_block + 2;
     let mut setup = validator::testonly::Setup::from_spec(rng, spec);
-    setup.push_blocks(rng, EXCHANGED_STATE_COUNT);
+    setup.push_blocks_v1(rng, EXCHANGED_STATE_COUNT);
     let cfgs = testonly::new_configs(rng, &setup, gossip_peers);
     scope::run!(ctx, |ctx, s| async {
         let mut nodes = vec![];
@@ -144,7 +144,7 @@ async fn test_switching_on_nodes() {
     // because we spawn the nodes gradually and we want the network
     // to be connected at all times.
     let cfgs = testonly::new_configs(rng, &setup, setup.validator_keys.len());
-    setup.push_blocks(rng, cfgs.len());
+    setup.push_blocks_v1(rng, cfgs.len());
     scope::run!(ctx, |ctx, s| async {
         let mut nodes = vec![];
         for (i, mut cfg) in cfgs.into_iter().enumerate() {
@@ -199,7 +199,7 @@ async fn test_switching_off_nodes() {
     // because we spawn the nodes gradually and we want the network
     // to be connected at all times.
     let cfgs = testonly::new_configs(rng, &setup, setup.validator_keys.len());
-    setup.push_blocks(rng, cfgs.len());
+    setup.push_blocks_v1(rng, cfgs.len());
     scope::run!(ctx, |ctx, s| async {
         let mut nodes = vec![];
         for (i, mut cfg) in cfgs.into_iter().enumerate() {
@@ -256,7 +256,7 @@ async fn test_different_first_block() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = validator::testonly::Setup::new(rng, 4);
-    setup.push_blocks(rng, 10);
+    setup.push_blocks_v1(rng, 10);
     // It is important that all nodes will connect to each other,
     // because we spawn the nodes gradually and we want the network
     // to be connected at all times.
@@ -322,7 +322,7 @@ async fn test_sidechannel_sync() {
     let mut spec = validator::testonly::SetupSpec::new(rng, 2);
     spec.first_block = spec.first_pregenesis_block + 2;
     let mut setup = validator::testonly::Setup::from_spec(rng, spec);
-    setup.push_blocks(rng, 10);
+    setup.push_blocks_v1(rng, 10);
     let cfgs = testonly::new_configs(rng, &setup, 1);
     scope::run!(ctx, |ctx, s| async {
         let mut stores = vec![];
@@ -407,7 +407,7 @@ async fn test_syncing_without_pregenesis_support() {
     let mut spec = validator::testonly::SetupSpec::new(rng, 2);
     spec.first_block = spec.first_pregenesis_block + 2;
     let mut setup = validator::testonly::Setup::from_spec(rng, spec);
-    setup.push_blocks(rng, 6);
+    setup.push_blocks_v1(rng, 6);
     let mut cfgs = testonly::new_configs(rng, &setup, 1);
     cfgs[1].enable_pregenesis = false;
     for cfg in &mut cfgs {
