+++ node/libs/storage/src/tests.rs
@@ -13,7 +13,7 @@ async fn test_inmemory_block_store() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = Setup::new(rng, 3);
-    setup.push_blocks(rng, 5);
+    setup.push_blocks_v1(rng, 5);
 
     let store = &testonly::in_memory::BlockStore::new(
         &setup,
@@ -74,7 +74,7 @@ async fn test_get_not_cached_block() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = Setup::new(rng, 1);
-    setup.push_blocks(rng, block_store::CACHE_CAPACITY + 5);
+    setup.push_blocks_v1(rng, block_store::CACHE_CAPACITY + 5);
     scope::run!(ctx, |ctx, s| async {
         let store = TestMemoryStorage::new(ctx, &setup).await;
         s.spawn_bg(store.runner.run(ctx));
@@ -109,7 +109,7 @@ async fn test_state_updates() {
     let ctx = &ctx::test_root(&ctx::RealClock);
     let rng = &mut ctx.rng();
     let mut setup = Setup::new(rng, 1);
-    setup.push_blocks(rng, 5);
+    setup.push_blocks_v1(rng, 5);
     // Create store with non-trivial first block.
     let first_block = &setup.blocks[2];
     let store =
